URBANICITY OVERHAUL
based on some initial reports of the weights, we are doing a very poor job of predicting things that should be associated with urbanicity for specific clusters (namely: 125,47,320,97,226,39,509, 476, 467, 537, 497). Given that PCA is dimension reduction, even though it was accounting for ~80% of the var for our four explanatory covar, it appears that 20% was more important than I thought. As a result, am going to include all four covars as latent variables for urbanicity. Althoguh they are collinear, the L1 regularization of glmnet and RF and other covar shrinkage of the various ML algorithm should allow that to not be too much of an issue (particularly because we are concerned about a predictive prob and not a beta where the collinearity would cause beta to be misspec). This should help to resolve some of the underfitting